getROC {demopack}	R Documentation

Compute the value of ROC for markers or compare the values of ROC by using two methods

Description

Compute the value of ROC for markers or compare the values of ROC by using two methods

Usage

getROC(x)

Arguments

x
x is a data.frame.When x is a data.frame with four columns,the first column is the marker expression of the first method.The second column is the marker expression of the second method.The third column is the illustration of two methods.The forth column is the labels of markers When x is a data.frame with two columns,the first column is the marker expression of the first method.The second column is the labels of markers

Value

When x is a data.frame with four columns,the result is two figures of ROC and the values of ROC by using two methods.When x is a data.frame with two columns,the result is one figure of ROC and the value of ROC

Examples

get.ROC(data)#data is a data.frame

[Package demopack version 0.1.0 Index]